chore: move hot next function from runner to runner factory
